# backend/src/main/java/io/kellermann/bigcontainers/controller/ReturnStockRequest.java
package io.kellermann.bigcontainers.controller;

import jakarta.validation.constraints.NotNull;
import java.math.BigDecimal;
import java.util.UUID;

/** Request body for {@code POST /api/v1/asset-models/{assetModelId}/consumable-stock/return}. */
public record ReturnStockRequest(
        UUID containerAssetId, UUID locationId, @NotNull BigDecimal quantity, String note, UUID eventReferenceId) {}