# frontend/packages/api-client/openapi/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin application server"
    }
  ],
  "paths": {
...
    "/api/v1/asset-models/{assetModelId}/consumable-stock/transfer": {
      "post": {
        "tags": ["consumable-stock-controller"],
        "operationId": "transfer",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferStockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StockTransferResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset-models/{assetModelId}/consumable-stock/return": {
      "post": {
        "tags": ["consumable-stock-controller"],
        "operationId": "returnStock",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnStockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumableStockResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset-models/{assetModelId}/consumable-stock/receive": {
      "post": {
        "tags": ["consumable-stock-controller"],
        "operationId": "receive",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiveStockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumableStockResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset-models/{assetModelId}/consumable-stock/issue": {
      "post": {
        "tags": ["consumable-stock-controller"],
        "operationId": "issue",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueStockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumableStockResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset-models/{assetModelId}/consumable-stock/consume": {
      "post": {
        "tags": ["consumable-stock-controller"],
        "operationId": "consume",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeStockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumableStockResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset-models/{assetModelId}/consumable-stock/adjust": {
      "post": {
        "tags": ["consumable-stock-controller"],
        "operationId": "adjust",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustStockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumableStockResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/locations/{locationId}/consumable-stock": {
      "get": {
        "tags": ["consumable-stock-controller"],
        "operationId": "listAtLocation",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsumableStockResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consumable-stock/{balanceId}": {
      "get": {
        "tags": ["consumable-stock-controller"],
        "operationId": "get_2",
        "parameters": [
          {
            "name": "balanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumableStockResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consumable-stock/{balanceId}/movements": {
      "get": {
        "tags": ["consumable-stock-controller"],
        "operationId": "movements",
        "parameters": [
          {
            "name": "balanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockMovementResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assets/{assetId}/consumable-stock": {
      "get": {
        "tags": ["consumable-stock-controller"],
        "operationId": "listAtContainer",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsumableStockResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/asset-models/{assetModelId}/consumable-stock": {
      "get": {
        "tags": ["consumable-stock-controller"],
        "operationId": "listByModel",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsumableStockResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset-models/{assetModelId}/consumable-stock/summary": {
      "get": {
        "tags": ["consumable-stock-controller"],
        "operationId": "summary",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetModelStockSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TransferStockRequest": {
        "type": "object",
        "properties": {
          "sourceContainerAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceLocationId": {
            "type": "string",
            "format": "uuid"
          },
          "destinationContainerAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "destinationLocationId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number"
          },
          "note": {
            "type": "string"
          }
        },
        "required": ["quantity"]
      },
      "ConsumableStockResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "assetModelId": {
            "type": "string",
            "format": "uuid"
          },
          "assetModelName": {
            "type": "string"
          },
          "stockUnitLabel": {
            "type": "string"
          },
          "containerAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "containerAssetDisplayName": {
            "type": "string"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "locationDisplayName": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StockTransferResponse": {
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/ConsumableStockResponse"
          },
          "destination": {
            "$ref": "#/components/schemas/ConsumableStockResponse"
          }
        }
      },
      "ReturnStockRequest": {
        "type": "object",
        "properties": {
          "containerAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number"
          },
          "note": {
            "type": "string"
          },
          "eventReferenceId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["quantity"]
      },
      "ReceiveStockRequest": {
        "type": "object",
        "properties": {
          "containerAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number"
          },
          "note": {
            "type": "string"
          }
        },
        "required": ["quantity"]
      },
      "IssueStockRequest": {
        "type": "object",
        "properties": {
          "containerAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number"
          },
          "note": {
            "type": "string"
          },
          "eventReferenceId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["quantity"]
      },
      "ConsumeStockRequest": {
        "type": "object",
        "properties": {
          "containerAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number"
          },
          "note": {
            "type": "string"
          }
        },
        "required": ["quantity"]
      },
      "AdjustStockRequest": {
        "type": "object",
        "properties": {
          "containerAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "delta": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "RECEIPT",
              "TRANSFER",
              "EVENT_ISSUE",
              "EVENT_RETURN",
              "CONSUMPTION",
              "AUDIT_ADJUSTMENT",
              "MANUAL_ADJUSTMENT"
            ]
          },
          "reason": {
            "type": "string"
          },
          "auditReferenceId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["delta", "type"]
      },
...
      "StockMovementResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "consumableStockBalanceId": {
            "type": "string",
            "format": "uuid"
          },
          "quantityDelta": {
            "type": "number"
          },
          "resultingQuantity": {
            "type": "number"
          },
          "stockUnitLabel": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "enum": [
              "RECEIPT",
              "TRANSFER",
              "EVENT_ISSUE",
              "EVENT_RETURN",
              "CONSUMPTION",
              "AUDIT_ADJUSTMENT",
              "MANUAL_ADJUSTMENT"
            ]
          },
          "actorUserId": {
            "type": "string",
            "format": "uuid"
          },
          "note": {
            "type": "string"
          },
          "transferGroupId": {
            "type": "string",
            "format": "uuid"
          },
          "eventReferenceId": {
            "type": "string",
            "format": "uuid"
          },
          "auditReferenceId": {
            "type": "string",
            "format": "uuid"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AssetModelStockSummaryResponse": {
        "type": "object",
        "properties": {
          "assetModelId": {
            "type": "string",
            "format": "uuid"
          },
          "assetModelName": {
            "type": "string"
          },
          "stockUnitLabel": {
            "type": "string"
          },
          "totalQuantity": {
            "type": "number"
          },
          "lowStockThreshold": {
            "type": "number"
          },
          "lowStock": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}